backend/requirements.txt
Flask>=2.2.0
flask-cors>=3.0.10
flask-sqlalchemy>=3.0.2
flask-migrate>=4.0.4
psycopg2-binary>=2.9.9
python-dotenv>=0.19.0
bcrypt>=4.0.1
pyotp>=2.9.0
tenacity>=8.2.3
requests>=2.31.0
SQLAlchemy>=1.4.47
alembic>=1.13.1
tenacity>=9.0.0